requirements.txt
Flask
torch
transformers
faiss-cpu
Pillow
google-generativeai
tqdm
numpy
